feat(react-ui): rename notification folder
