+++ libs/react-ui/src/lib/notification/Notification.scss
